Add files via upload
This simulates spread arbitrage among various exchanges or products. This is one-stop service of trading. Once incorporating data, this program automatically calculates spread among pairs, and optimizes the strategy for each pair's learning period (in the sample), and then applying the parameter to out of sample data to get the trading result for each pair.
A further step would be linking API and getting data from it, and enabling trade on it in real time manner.